infra/cdk/requirements.txt
aws_cdk.core
aws_cdk.aws_s3_deployment
aws_cdk.aws_s3
aws_cdk.aws_route53
aws_cdk.aws_route53_targets
aws_cdk.aws_cloudfront
